requirements.txt
networkx
spacy
pyvis
faiss-cpu
sentence-transformers
huggingface_hub
openai
bert-score
